Add Top Module
Wrote the module for the top half of the holder. Re-uses a lot of code
from the bottom module, so it may be a good idea to factor that out
of both. There is also a hard-coded expansion factor on the recesses
between the squares that cap off each compartment. Hopefully that will
help make a better fit when printed. In the future, it should be
parameterized or removed.