Add expansion fudge factor
A factor has been added that allows for slight adjustment of the compartment
cap size on the top half of the box. If the top and bottom half are designed
to fit "perfectly", then expansion of the plastic during the printing may
cause the fit to be too tight, so that the box won't close. The fudge factor
(named "expf") lets the user account for this by slightly increasing the
spacing in between the squares on the top half. For expf=0.1, for example,
the spacing will be 10% wider than the wall thickness.